backend/requirements.txt
asgiref==3.7.2
Django==3.2.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2
dj-database-url
python-decouple
djangorestframework==3.14.0
pillow==10.0.0
djoser==2.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
django-templated-mail==1.1.1
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
pycparser==2.21
pyjwt==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2 
urllib3==2.0.4
django-filter==23.2
drf-base64==2.0
isort==5.12.0
python-dotenv==1.0.0